Fixed issue displaying array types #829
Repost: The data type displayed in the Model block currently only
indicates an array when a $ref is used. These changes will fix the
Model section to include the array wrapper around primitive data types
(references were already working). This also fixes a bug where the data
type is not displayed in the Model Schema block.