Fix a null pointer in OperationView.js in headers
The check for "Download" content type in OperationView.js does not
determine if the header is available before checking the regular
expression. If the header is not available, then the operation will
fail and the behavior will resemble:

https://github.com/swagger-api/swagger-ui/issues/1109

This issue appears to extend to other content types, including
application/xml, or any not explicitly defined before the
"Download" conditional. Simply checking for a null value before
testing the regular expression is enough to get things to work
properly again.