Respect the name of the api key supplied in spec
The code used to always use "api_key" as the name of the api key query parameter, but the spec has a provision for supplying a different name. This will now use the name of the key supplied in the spec (if one exists), and will fall back to api_key if none is specified.