Allow swaggerRequestHeaders as url parameter to index.html
It's nice if you can have a running swagger-ui somewhere and just url-link to a swagger-spec
you have in source control. That would look like this:

https://acme.com/swaggerui/?url=../bitbucket/repos/spec/browse/spec.json?raw

Sadly this breaks for bitbucket and similar tools, because swagger-ui requests the spec as 'application/json' by default.

The new url-parameter 'swaggerRequestHeaders' solves this problem by using the existing
configuration option 'swaggerRequestHeaders'. For example like so:

https://acme.com/swaggerui/?swaggerRequestHeaders=text/html&url=../bitbucket/repos/spec/browse/spec.json?raw