Added support for passing through a basePath override.
For cases where the API is being served from behind a reverse proxy, added
support for setting the basePath of the SwaggerClient object using an option
in SwaggerUI construction.

Since this needs to happen after the client has parsed the specification,
the update is hooked into the "render" method, despite that sounding kind
of weird when you write it down.