Custom Header support for OAuth2 Authorization flows
While the regular Swagger UI configuration permits intercepting and
decorating requests using the `requestInterceptor` option, the same
is not available for the OAuth2 Authorization requests. Thus it
is not possible to add common CSRF protection headers required by
some API's.

As the OAuth2 configuration seems to have a slightly different signature
than the usual swagger UI config, this patch adds an `additionalHeaders`
field (similar to `additionalQueryStringParams`) on the initOAuth2 configuration
input, which is attached to all OAuth2 based API requests.