Update response-body.jsx
In internet explorer, the "Download File" button is shown as a link, but when clicked nothing happens.
Right clicking the link and selecting Save as... will ignore file name, and only display the blob name.
The file has the proper content though. 

Internet explorer uses window.navigator.msSaveOrOpenBlob to save files.

Branching in this file, will allow proper functionality to the link, while having no breaking changes. 

This solves issues:
https://github.com/swagger-api/swagger-ui/issues/3364

The error can be seen on this :
http://swashbuckletest.azurewebsites.net/swagger/ui/index#/Image/Image_Put

This way, the internet explorer little window in the middle of the bottom of the screen will pop up, with the proper filename.