Enable Commonmark table extension
Commonmark itself does not support tables, but instead delegates that kind of functionality to extensions. The Remarkable configuration preset "commonmark" is a strict interpretation of the spec. This enables the table extension which brings Swagger UI back in line with the 2.x line which had this same extension enabled for Markdown. Other tools such as Apiary that support OAS 3/Commonmark are also enabling this extension, so there is precedence for doing this outside of Swagger UI as well.