Disable the validation badge for those who do not want it
When running swaggerui on Docker with a hostname, which is not
localhost, one gets a badge that says that the validation URI cannot be
accessed. Which is correct, because the URI is most likely to be on the
localhost. The check is done when the instance itself runs on 127.0.0.1
or localhost, but it ignores the VALIDATION_URI settings. The
documentation tells us when the VALIDATION_URI is set to `null` the
validation isn't done. This isn't true, the code in
`src/core/components/online-validator-badge.jsx` sets a default
`validatorUrl` when this is undefined.

This changeset changes this, to not display the validation badge if the
validation URI is 127.0.0.1, localhost, none or empty.

There could also be another option, which would be to introduce a new
configuration parameter: `validationEnabled`, which would tell is that
we need validation. This would eliminate the need to check if the host
is on localhost and would allow even more fine-grained control over when
and where the validation can be done. I've chosen not to go that route
because I'm unfamiliar with the code base and this seemed to be the
smallest change to accomplish my needs for a developing OpenAPI specs
with Docker without seeing a validation error badge.

Signed-off-by: Wesley Schwengle <wesley@opndev.io>