Update how the JavaScript run function is invoked in oauth2-redirect.html
We're embedding Swagger UI in an application that defines a content security policy &mdash; see [an introduction on MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP) if necessary.

One issue we're facing is that the `oauth2-redirect.html` file uses an inline event handler, `<body onload="run()">`, which is not supported by Chrome or Edge.
The official Chrome docs mention that those should be rewritten using the `addEventListener` method.
They also recommend to use the `DOMContentLoaded` event on the document rather than the `load` event on the window, as the former "generally triggers more quickly".
See <https://developer.chrome.com/extensions/contentSecurityPolicy#:~:text=The%20inline%20event%20handler,more%20quickly>

Because of this, we currently have to maintain a slightly modified copy of `oauth2-redirect.html` file to which we apply this change.
We're hoping that you'll accept that change to be made upstream.
We've tested it on Chrome, the new Edge, and Firefox.

Related issue: <https://github.com/swagger-api/swagger-ui/issues/5720>