fix: preserve multipart file part position in requestData
when providing type string format binary aka file position is not preserved,
because it is not registered on component mount like all others do.
So the state is later on corrupted,
because when the file is set and so a new property will be added to requestData Map in state.
This results in beeing the last property.
since all other properties' components are registering their value on component mount,
the oder is preserved for those.

Signed-off-by: mathis-m <mathis.michel@outlook.de>